PuzzleIOI with masks

+++ scripts/preprocess/copy_ioi_from_cluster.py
@@ -44,7 +44,7 @@ def copy_all(src_dir, tgt_dir):
 
     src_meshs = glob(join(src_dir, "meshes", "*000001.obj"))
 
-    if len(src_meshs) == 1 and (not os.path.exists(join(tgt_dir, "scan.obj"))):
+    if len(src_meshs) == 1:
 
         for type in ["images"]:
             mkdir(join(tgt_dir, type))
@@ -53,16 +53,25 @@ def copy_all(src_dir, tgt_dir):
             if not os.path.exists(tgt_img):
                 shutil.copyfile(src_img, tgt_img)
 
-        src_meshs = glob(join(src_dir, "meshes", "*000001.obj"))
-
-        src_mesh = src_meshs[0]
-
+        src_mesh = glob(join(src_dir, "meshes", "*000001.obj"))[0]
+        src_mtl = glob(join(src_dir, "meshes", "*000001.mtl"))[0]
+        src_jpg = glob(join(src_dir, "meshes", "*000001.jpg"))[0]
         src_cam = glob(join(src_dir, "meshes", "*.csd"))[0]
+
         tgt_mesh = join(tgt_dir, "scan.obj")
         tgt_cam = join(tgt_dir, "camera.csd")
+        tgt_mtl = join(tgt_dir, os.path.basename(src_mtl))
+        tgt_jpg = join(tgt_dir, os.path.basename(src_jpg))
 
-        shutil.copyfile(src_mesh, tgt_mesh)
-        shutil.copyfile(src_cam, tgt_cam)
+        if not os.path.exists(tgt_mesh):
+            shutil.copyfile(src_mesh, tgt_mesh)
+        if not os.path.exists(tgt_cam):
+            shutil.copyfile(src_cam, tgt_cam)
+        if not os.path.exists(tgt_mtl):
+            shutil.copyfile(src_mtl, tgt_mtl)
+        if not os.path.exists(tgt_jpg):
+            shutil.copyfile(src_jpg, tgt_jpg)
+
 
 def copy_dataset(subject, src_path, tgt_path):
 
@@ -91,22 +100,21 @@ def copy_dataset(subject, src_path, tgt_path):
             ][0]
             src_outfit_dir = join(motion_dir, outfit_name)
             tgt_outfit_dir = join(tgt_path, "fitting", person_id, tgt_outfit_name)
-            
+
             copy_all(src_outfit_dir, tgt_outfit_dir)
 
         all_front_paths = sorted(glob(join(tgt_path, "fitting", person_id, "*/images/07_C.jpg")))
-        
+
         try:
-            all_full_imgs = [
-                read_image(img_path)
-                for img_path in all_front_paths
-            ]
-            cv2.imwrite(
-                join(tgt_path, "fitting", person_id, "full.jpg"),
-                make_grid(all_full_imgs, nrow=5).permute(1, 2, 0).numpy()[:, :, ::-1]
-            )
+            all_full_imgs = [read_image(img_path) for img_path in all_front_paths]
+            full_path = join(tgt_path, "fitting", person_id, "full.jpg")
+            if not os.path.exists(full_path):
+                cv2.imwrite(
+                    full_path,
+                    make_grid(all_full_imgs, nrow=5).permute(1, 2, 0).numpy()[:, :, ::-1]
+                )
         except:
-            myerror(f"cannot read {all_front_paths}")    
+            myerror(f"cannot read {all_front_paths}")
 
     # print(f"{person_id} is done!")
 
@@ -121,7 +129,7 @@ if __name__ == "__main__":
 
     mkdir(join(args.tgt_path, "fitting"))
     subjects = sorted(os.listdir(args.src_path))
-    
+
     # subjects = ["DynamicClothCap_220426_03612_CG"]
 
     with Pool(processes=mp.cpu_count(), maxtasksperchild=1) as pool:
